+++ packages/gui/src/components/inputs/Dimension/Input.tsx
@@ -1,23 +1,34 @@
 import * as React from 'react'
 import {
   AbsoluteLengthUnits,
+  CalcOperand,
+  CSSFunctionCalc,
   CSSUnitValue,
+  Dimension,
   KeywordUnits,
   ThemeUnits,
+  UnitlessUnits,
 } from '../../../types/css'
-import { Label, Number, ThemeValue, UnitSelect, ValueSelect } from '../../primitives'
+import {
+  Label,
+  Number,
+  ThemeValue,
+  UnitSelect,
+  ValueSelect,
+} from '../../primitives'
 import { reducer } from './reducer'
 import { State } from './types'
 import { EditorProps } from '../../../types/editor'
 import { UnitConversions } from '../../../lib/convert'
 import { compact, kebabCase } from 'lodash-es'
+import { CalcInput } from '../../primitives/CalcInput'
 
 // Mapping of units to [min, max] tuple
 type UnitRanges = Record<string, [min: number, max: number]>
 // Mapping of units to steps
 type UnitSteps = Record<string, number>
 
-export interface DimensionInputProps extends EditorProps<CSSUnitValue> {
+export interface DimensionInputProps extends EditorProps<Dimension> {
   label?: string
   range?: UnitRanges
   steps?: UnitSteps
@@ -41,17 +52,19 @@ export const DimensionInput = ({
 }: DimensionInputProps) => {
   const id = `${React.useId()}-${kebabCase(label)}`
   const [state, dispatch] = React.useReducer(reducer, {
-    value: value?.value || 0,
-    unit: value?.unit || units[0] || AbsoluteLengthUnits.Px,
-    themeId: value?.themeId,
+    value: (value as CSSUnitValue)?.value || 0,
+    unit: (value as CSSUnitValue)?.unit || units[0] || AbsoluteLengthUnits.Px,
+    themeId: (value as CSSUnitValue)?.themeId,
     key: 0,
   } as State)
   React.useEffect(() => {
+    if ((value as CSSFunctionCalc)?.type === 'calc') return
+    const unitValue = value as CSSUnitValue
     if (
       // Only want to call on change when the value differs
-      state.value !== value?.value ||
-      state.unit !== value?.unit ||
-      state.themeId !== value?.themeId
+      state.value !== unitValue?.value ||
+      state.unit !== unitValue?.unit ||
+      state.themeId !== unitValue?.themeId
     ) {
       const newValue: CSSUnitValue = {
         value: state.value,
@@ -68,6 +81,7 @@ export const DimensionInput = ({
     themeValues.length > 0 && 'theme',
     ...units,
     keywords.length > 0 && 'keyword',
+    UnitlessUnits.Calc,
   ])
 
   return (
@@ -104,9 +118,21 @@ export const DimensionInput = ({
                 type: 'CHANGED_INPUT_TO_THEME_VALUE',
                 value: themeValue?.value ?? 0,
                 unit: (themeValue?.unit as any) ?? 'px',
-                themeId: themeValue.id
+                themeId: themeValue.id,
               })
             }}
+            themeValues={themeValues}
+          />
+        ) : state.unit === UnitlessUnits.Calc ? (
+          <CalcInput
+            units={allUnits}
+            onChange={onChange}
+            //@ts-ignore
+            value={value}
+            label={label}
+            steps={steps}
+            range={range}
+            conversions={conversions}
             themeValues={themeValues}
           />
         ) : (
@@ -131,6 +157,37 @@ export const DimensionInput = ({
           onChange={(e: React.ChangeEvent<HTMLSelectElement>) => {
             const newUnit = e.target.value
 
+            if (newUnit === KeywordUnits.Keyword) {
+              dispatch({
+                type: 'CHANGED_INPUT_VALUE',
+                value: keywords[0],
+              })
+            }
+
+            if (newUnit === UnitlessUnits.Calc) {
+              onChange({
+                arguments: {
+                  valueX: value as CSSUnitValue,
+                  valueY: { value: 1, unit: 'px' },
+                  operand: CalcOperand.Plus,
+                },
+                type: 'calc',
+              })
+            }
+            if (
+              state.unit === UnitlessUnits.Calc &&
+              newUnit !== UnitlessUnits.Calc
+            ) {
+              const unitValue = (value as CSSFunctionCalc).arguments.valueX
+                .value
+
+              onChange({ value: unitValue, unit: newUnit })
+              dispatch({
+                value: unitValue,
+                type: 'CHANGED_INPUT_VALUE',
+              })
+            }
+
             if (newUnit === ThemeUnits.Theme) {
               const themeValue = themeValues?.[0]
               return dispatch({
@@ -138,13 +195,6 @@ export const DimensionInput = ({
                 value: themeValue?.value ?? 0,
                 unit: (themeValue?.unit as any) ?? 'px',
                 themeId: themeValue?.id,
-              })
-            }
-
-            if (newUnit === KeywordUnits.Keyword) {
-              dispatch({
-                type: 'CHANGED_INPUT_VALUE',
-                value: keywords[0],
               })
             }