+++ packages/gui/src/components/inputs/Dimension/types.ts
@@ -1,4 +1,5 @@
 import { UnitConversions, UnitSteps } from '../../../lib'
+import { CSSFunctionCalc } from '../../../types/css'
 
 export type State = {
   key: number